backend/pyproject.toml
[project]
name = "search"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "aiofiles>=25.1.0",
    "click>=8.3.0",
    "fastapi[standard]>=0.121.0",
    "httpx>=0.28.1",
    "huggingface-hub>=0.36.0",
    "loguru>=0.7.3",
    "numpy>=2.3.4",
    "opencv-python>=4.11.0.86",
    "pandas>=2.3.3",
    "pillow>=12.0.0",
    "pydantic>=2.12.4",
    "pydantic-settings>=2.11.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.20",
    "qdrant-client>=1.15.1",
    "rank-bm25>=0.2.2",
    "requests>=2.32.5",
    "scikit-image>=0.25.2",
    "sentence-transformers>=3.3.1",
    "torch>=2.9.0",
    "torchvision>=0.24.0",
    "tqdm>=4.67.1",
    "transformers>=4.57.1",
    "ultralytics>=8.3.225",
    "uvicorn[standard]>=0.38.0",
]
